refactor(add-data): split AddDataDialog into per-source components (#350)
Break the ~3000-line AddDataDialog god-component into a thin dialog shell
plus one self-contained component per data source (XYZ, WMS, WFS, WMTS,
GPX, delimited text, MBTiles, ArcGIS, PostgreSQL, deck.gl, video). Each
source now owns its own form state and submit logic and reads shared
services (store, map controller, submit-in-progress, close handling) from
the shell via context.

Pure helpers, constants, and type aliases move into add-data/ modules. The
Martin connection is lifted into a useMartinConnection hook in the shell so
a running server still survives source remounts, preserving the original
behavior. No behavior change.

Adds tests for the extracted pure helpers.
